Lab3 done - add JPA database persistance

+++ src/main/java/mk/ukim/finki/wp/lab/service/impl/BookReservationServiceImpl.java
@@ -29,7 +29,7 @@ public class BookReservationServiceImpl implements BookReservationService {
                 bookTitle,
                 readerName,
                 readerAddress,
-                (long) numberOfCopies
+                numberOfCopies
         );
 
         return bookReservationRepository.save(bookReservation);
